+++ CHANGELOG.md
@@ -2,6 +2,15 @@
 
 All notable changes to this project will be documented in this file.
 
+## \[v0.3.0\]
+### Updated ASHA
+* Internal algorithm of the AHSA implementation use iteration unit instead of epoch.
+* Minor bug fixes.
+
+## \[v0.2.0\]
+### Updated
+* Added get_progress() API to provide current progresses.
+
 ## \[v0.1.0\]
 ### Added
 * Initial version of hpopt.